refactor: Remove redundant font handling in layout
The fonts.ts file and its usage in layout.tsx have been deleted due to redundancy. We used to import fonts from the 'next/font/google' package and assign them to variables in fonts.ts, then import these variables in the layout.tsx. However, we concluded that this can be streamlined by directly adding the required font-families as global CSS properties. This change will contribute to cleaner and simpler code.